- fixed compile warning on OpenBSD
git-svn-id: https://openser.svn.sourceforge.net/svnroot/openser/trunk@1399 689a6050-402a-0410-94f2-e92a70836424

+++ modules_k/acc/acc.c
@@ -251,7 +251,7 @@ int acc_log_request( struct sip_msg *rq)
 	*(p++) = 0;
 
 	LOG(log_level, "%.*stimestamp=%lu%s",
-		acc_env.text.len, acc_env.text.s, acc_env.ts, log_msg);
+		acc_env.text.len, acc_env.text.s,(unsigned long) acc_env.ts, log_msg);
 
 	return 1;
 }
